Add filth_type API
Certain enemies (trash_tire, porcupine, stoneboss, and wolf) spawn
hitboxes with the a different hard coded filth type.
As far as I can tell this value is not used anywhere, either in-engine
or in any scripts, so it should be safe to change to use the entity's
filth_type for consistency.